+++ Changelog.md
@@ -1,3 +1,10 @@
+### 0.9
+* Added "Alternative Advantage" which allows you to specify a roll formula to add/subtract to a roll for advantage/disadvantage instead of rolling 2 dice.
+  - The roll is evaluated against the actor/items roll data so expressions like "1d4 + @abilities.str.mod" will work, or just a number like 2
+
+### 0.8
+* Allow item recharge to be silent or generate a chat message.
+
 ### 0.7
 * Fix for not working in non-english environments